api: phase 2c — relocate pixel_to_gripper_point into rig_laserline_device
Move the free function `pixel_to_gripper_point` out of the facade crate
root (its only original implementation) into a new `geometry` module of
`vision-calibration-pipeline`'s `rig_laserline_device`. The facade now
re-exports it at `vision_calibration::rig_laserline_device::pixel_to_gripper_point`
and keeps a `#[deprecated]` crate-root alias for one release. lib.rs is now
a pure facade with no original function bodies.
